+++ migrations/versions/3e08dd518f9c_add_my_cool_nrel_atb_tables.py
@@ -1,8 +1,8 @@
 """Add my cool NREL ATB tables
 
-Revision ID: e2e6ca8a3a4e
+Revision ID: 3e08dd518f9c
 Revises: 172898b30824
-Create Date: 2024-04-23 15:41:45.302449
+Create Date: 2024-04-23 16:03:19.347486
 
 """
 from alembic import op
@@ -10,7 +10,7 @@ import sqlalchemy as sa
 
 
 # revision identifiers, used by Alembic.
-revision = 'e2e6ca8a3a4e'
+revision = '3e08dd518f9c'
 down_revision = '172898b30824'
 branch_labels = None
 depends_on = None
@@ -22,7 +22,7 @@ def upgrade() -> None:
     sa.Column('report_year', sa.Integer(), nullable=False, comment='Four-digit year in which the data was reported.'),
     sa.Column('core_metric_case', sa.Enum('Market', 'R&D'), nullable=False, comment="NREL's financial assumption cases. There are two cases which effect project finanical assumptions: R&D Only Case and Market + Policies Case. R&D Only includes only projected R&D improvements while Market + Policy case includes policy and tax incentives. https://atb.nrel.gov/electricity/2023/financial_cases_&_methods"),
     sa.Column('core_metric_variable_year', sa.Integer(), nullable=False, comment='The year of the projected value.'),
-    sa.Column('technology_description', sa.Enum('DistributedWind', 'OffShoreWind', 'Coal_Retrofits', 'CSP', 'Pumped Storage Hydropower', 'LandbasedWind', 'Residential Battery Storage', 'CommPV', 'Commercial Battery Storage', 'AEO', 'ResPV', 'Nuclear', 'Coal_FE', 'Biopower', 'NaturalGas_FE', 'Geothermal', 'Utility-Scale PV-Plus-Battery', 'Hydropower', 'NaturalGas_Retrofits', 'Utility-Scale Battery Storage', 'UtilityPV'), nullable=False, comment='High level description of the technology used by the generator to produce electricity.'),
+    sa.Column('technology_description', sa.Enum('Commercial Battery Storage', 'CommPV', 'Utility-Scale Battery Storage', 'DistributedWind', 'ResPV', 'Utility-Scale PV-Plus-Battery', 'LandbasedWind', 'Coal_Retrofits', 'Hydropower', 'Nuclear', 'OffShoreWind', 'Geothermal', 'NaturalGas_Retrofits', 'UtilityPV', 'Coal_FE', 'Residential Battery Storage', 'Biopower', 'Pumped Storage Hydropower', 'CSP', 'NaturalGas_FE', 'AEO'), nullable=False, comment='High level description of the technology used by the generator to produce electricity.'),
     sa.Column('inflation_rate', sa.Float(), nullable=True, comment='Rate of inflation. All dollar values are given in 2021 USD, using the Consumer Price Index for All Urban Consumers for dollar year conversions where the source year dollars do not match 2021.'),
     sa.Column('interest_rate_during_construction_nominal', sa.Float(), nullable=True, comment='Also referred to as construction finance cost. Portion of all-in capital cost associated with construction period financing. It is a function of construction duration, capital fraction during construction, and interest during construction.'),
     sa.Column('interest_rate_calculated_real', sa.Float(), nullable=True, comment='Calculated real interest rate.'),
@@ -36,7 +36,7 @@ def upgrade() -> None:
     sa.Column('report_year', sa.Integer(), nullable=True, comment='Four-digit year in which the data was reported.'),
     sa.Column('core_metric_case', sa.Enum('Market', 'R&D'), nullable=True, comment="NREL's financial assumption cases. There are two cases which effect project finanical assumptions: R&D Only Case and Market + Policies Case. R&D Only includes only projected R&D improvements while Market + Policy case includes policy and tax incentives. https://atb.nrel.gov/electricity/2023/financial_cases_&_methods"),
     sa.Column('core_metric_variable_year', sa.Integer(), nullable=True, comment='The year of the projected value.'),
-    sa.Column('technology_description', sa.Enum('DistributedWind', 'OffShoreWind', 'Coal_Retrofits', 'CSP', 'Pumped Storage Hydropower', 'LandbasedWind', 'Residential Battery Storage', 'CommPV', 'Commercial Battery Storage', 'AEO', 'ResPV', 'Nuclear', 'Coal_FE', 'Biopower', 'NaturalGas_FE', 'Geothermal', 'Utility-Scale PV-Plus-Battery', 'Hydropower', 'NaturalGas_Retrofits', 'Utility-Scale Battery Storage', 'UtilityPV'), nullable=True, comment='High level description of the technology used by the generator to produce electricity.'),
+    sa.Column('technology_description', sa.Enum('Commercial Battery Storage', 'CommPV', 'Utility-Scale Battery Storage', 'DistributedWind', 'ResPV', 'Utility-Scale PV-Plus-Battery', 'LandbasedWind', 'Coal_Retrofits', 'Hydropower', 'Nuclear', 'OffShoreWind', 'Geothermal', 'NaturalGas_Retrofits', 'UtilityPV', 'Coal_FE', 'Residential Battery Storage', 'Biopower', 'Pumped Storage Hydropower', 'CSP', 'NaturalGas_FE', 'AEO'), nullable=True, comment='High level description of the technology used by the generator to produce electricity.'),
     sa.Column('scenario_atb', sa.Enum('Advanced', 'Moderate', 'Conservative'), nullable=True, comment='Technology innovation scenarios. https://atb.nrel.gov/electricity/2023/definitions#scenarios'),
     sa.Column('cost_recovery_period_years', sa.Integer(), nullable=True, comment='The period over which the initial capital investment to build a plant is recovered.'),
     sa.Column('capital_recovery_factor', sa.Float(), nullable=True, comment='Ratio of a constant annuity to the present value of receiving that annuity for a given length of time.'),
@@ -50,7 +50,7 @@ def upgrade() -> None:
     sa.Column('report_year', sa.Integer(), nullable=True, comment='Four-digit year in which the data was reported.'),
     sa.Column('core_metric_case', sa.Enum('Market', 'R&D'), nullable=True, comment="NREL's financial assumption cases. There are two cases which effect project finanical assumptions: R&D Only Case and Market + Policies Case. R&D Only includes only projected R&D improvements while Market + Policy case includes policy and tax incentives. https://atb.nrel.gov/electricity/2023/financial_cases_&_methods"),
     sa.Column('core_metric_variable_year', sa.Integer(), nullable=True, comment='The year of the projected value.'),
-    sa.Column('technology_description', sa.Enum('DistributedWind', 'OffShoreWind', 'Coal_Retrofits', 'CSP', 'Pumped Storage Hydropower', 'LandbasedWind', 'Residential Battery Storage', 'CommPV', 'Commercial Battery Storage', 'AEO', 'ResPV', 'Nuclear', 'Coal_FE', 'Biopower', 'NaturalGas_FE', 'Geothermal', 'Utility-Scale PV-Plus-Battery', 'Hydropower', 'NaturalGas_Retrofits', 'Utility-Scale Battery Storage', 'UtilityPV'), nullable=True, comment='High level description of the technology used by the generator to produce electricity.'),
+    sa.Column('technology_description', sa.Enum('Commercial Battery Storage', 'CommPV', 'Utility-Scale Battery Storage', 'DistributedWind', 'ResPV', 'Utility-Scale PV-Plus-Battery', 'LandbasedWind', 'Coal_Retrofits', 'Hydropower', 'Nuclear', 'OffShoreWind', 'Geothermal', 'NaturalGas_Retrofits', 'UtilityPV', 'Coal_FE', 'Residential Battery Storage', 'Biopower', 'Pumped Storage Hydropower', 'CSP', 'NaturalGas_FE', 'AEO'), nullable=True, comment='High level description of the technology used by the generator to produce electricity.'),
     sa.Column('cost_recovery_period_years', sa.Integer(), nullable=True, comment='The period over which the initial capital investment to build a plant is recovered.'),
     sa.Column('scenario_atb', sa.Enum('Advanced', 'Moderate', 'Conservative'), nullable=True, comment='Technology innovation scenarios. https://atb.nrel.gov/electricity/2023/definitions#scenarios'),
     sa.Column('technology_description_detail_1', sa.Text(), nullable=True, comment='Technology details indicate resource levels and specific technology subcategories.'),
@@ -68,11 +68,20 @@ def upgrade() -> None:
     sa.Column('opex_fixed_per_kw', sa.Float(), nullable=True, comment='Fixed operation and maintenance expenses. Annual expenditures to operate and maintain equipment that are not incurred on a per-unit-energy basis.'),
     sa.Column('opex_variable_per_mwh', sa.Float(), nullable=True, comment='Operation and maintenance costs incurred on a per-unit-energy basis.')
     )
+    op.create_table('core_nrelatb__yearly_technology_status',
+    sa.Column('report_year', sa.Integer(), nullable=True, comment='Four-digit year in which the data was reported.'),
+    sa.Column('technology_description', sa.Enum('Commercial Battery Storage', 'CommPV', 'Utility-Scale Battery Storage', 'DistributedWind', 'ResPV', 'Utility-Scale PV-Plus-Battery', 'LandbasedWind', 'Coal_Retrofits', 'Hydropower', 'Nuclear', 'OffShoreWind', 'Geothermal', 'NaturalGas_Retrofits', 'UtilityPV', 'Coal_FE', 'Residential Battery Storage', 'Biopower', 'Pumped Storage Hydropower', 'CSP', 'NaturalGas_FE', 'AEO'), nullable=True, comment='High level description of the technology used by the generator to produce electricity.'),
+    sa.Column('technology_description_detail_1', sa.Text(), nullable=True, comment='Technology details indicate resource levels and specific technology subcategories.'),
+    sa.Column('technology_description_detail_2', sa.Text(), nullable=True, comment='Technology details indicate resource levels and specific technology subcategories.'),
+    sa.Column('is_technology_mature', sa.Boolean(), nullable=True, comment='Indicator of whether the technology is mature.'),
+    sa.Column('is_default', sa.Boolean(), nullable=True, comment='Indicator of whether the technology is default.')
+    )
     # ### end Alembic commands ###
 
 
 def downgrade() -> None:
     # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('core_nrelatb__yearly_technology_status')
     op.drop_table('core_nrelatb__yearly_projections_by_technology_detail')
     op.drop_table('core_nrelatb__yearly_financial_cases_by_scenario')
     op.drop_table('core_nrelatb__yearly_financial_cases')